+++ CHANGELOG.md
@@ -6,3 +6,5 @@ All notable changes to this project will be documented in this file.
 - Multiple pages
 - Add page viewed statistics
 - Add clicked links statistics
+- Send link's click in Google Analytics
+- Add Google fonts selector for links, labels, separators and header